requirements.txt
requests>=2.31.0
gitpython>=3.1.40
click>=8.1.7
colorama>=0.4.6
PyYAML>=6.0.1
pyinstaller>=6.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
